lab 3

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\AboutController;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\BlogController;
 use App\Http\Controllers\ContactController;
 use App\Http\Controllers\GuestBookController;
@@ -10,6 +11,8 @@ use App\Http\Controllers\LearningController;
 use App\Http\Controllers\MainController;
 use App\Http\Controllers\PhotoController;
 use App\Http\Controllers\TestController;
+use App\Http\Controllers\SpyController;
+
 
 use Illuminate\Support\Facades\Route;
 
@@ -23,32 +26,39 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
+Route::get('/',                  [MainController::class,       'index'])    -> name('main.index');
+Route::get('/main',              [MainController::class,       'index'])    -> name('main.index');
+Route::get('/about',             [AboutController::class,      'index'])    -> name('about.index');
+Route::get('/contact',           [ContactController::class,    'index'])    -> name('contact.index');
+Route::get('/history',           [HistoryController::class,    'index'])    -> name('history.index');
+Route::get('/hobbies',           [HobbiesController::class,    'index'])    -> name('hobbies.index');
+Route::get('/learning',          [LearningController::class,   'index'])    -> name('learning.index');
+Route::get('/photo',             [PhotoController::class,      'index'])    -> name('photo.index');
+Route::get('/test',              [TestController::class,       'index'])    -> name('test.index');
 
-Route::get('/', [MainController::class, 'index'])->name('main.index');
-Route::get('/main', [MainController::class, 'index'])->name('main.index');
-Route::get('/about', [AboutController::class, 'index'])->name('about.index');
-Route::get('/contact', [ContactController::class, 'index'])->name('contact.index');
-Route::get('/history', [HistoryController::class, 'index'])->name('history.index');
-Route::get('/hobbies', [HobbiesController::class, 'index'])->name('hobbies.index');
-Route::get('/learning', [LearningController::class, 'index'])->name('learning.index');
-Route::get('/photo', [PhotoController::class, 'index'])->name('photo.index');
-Route::get('/test', [TestController::class, 'index'])->name('test.index');
+Route::post('/contact',          [ContactController::class,    'store'])    -> name('contact.store');
+Route::post('/test',             [TestController::class,       'store'])    -> name('test.store');
+Route::get('/test/table',        [TestController::class,       'table'])    -> name('test.table')       -> middleware('authorized');
 
-Route::post('/contact', [ContactController::class, 'store'])->name('contact.store');
-Route::post('/test', [TestController::class, 'store'])->name('test.store');
-Route::get('/test/table', [TestController::class, 'table'])->name('test.table');
+Route::get('/guestbook',         [GuestBookController::class,  'index'])    -> name('guestbook.index');
+Route::post('/guestbook',        [GuestBookController::class,  'store'])    -> name('guestbook.store');
+Route::get('/guestbook/edit',    [GuestBookController::class,  'edit'])     -> name('guestbook.edit')   -> middleware('admin');
+Route::post('/guestbook/update', [GuestBookController::class,  'update'])   -> name('guestbook.update') -> middleware('admin');
 
-Route::get('/guestbook', [GuestBookController::class, 'index'])->name('guestbook.index');
-Route::post('/guestbook', [GuestBookController::class, 'store'])->name('guestbook.store');
-Route::get('/guestbook/edit', [GuestBookController::class, 'edit'])->name('guestbook.edit');
-Route::post('/guestbook/update', [GuestBookController::class, 'update'])->name('guestbook.update');
+Route::get('/blog',              [BlogController::class,       'index'])    -> name('blog.index');
+Route::get('/blog/create',       [BlogController::class,       'create'])   -> name('blog.create')      -> middleware('authorized');
+Route::post('/blog',             [BlogController::class,       'store'])    -> name('blog.store')       -> middleware('authorized');
+Route::get('/blog/file_upload',  [BlogController::class,       'file_upload']) -> name('blog.file_upload')    -> middleware('admin');
+Route::post('/blog/file_upload', [BlogController::class,       'file_upload_update']) -> name('blog.file_upload_update') -> middleware('admin');
+Route::get('/blog/{id}',         [BlogController::class,       'show'])     -> name('blog.show');
+Route::get('/blog/{id}/edit',    [BlogController::class,       'edit'])     -> name('blog.edit')        -> middleware('authorized');
+Route::post('/blog/{id}',        [BlogController::class,       'update'])   -> name('blog.update')      -> middleware('authorized');
+Route::delete('/blog/{id}',      [BlogController::class,       'destroy'])  -> name('blog.destroy')     -> middleware('authorized');
+      
+Route::get('/login',             [AuthController::class,       'login_form']) -> name('auth.login_form');
+Route::post('/login',            [AuthController::class,       'login'])    -> name('auth.login');
+Route::get('/registration',      [AuthController::class,       'registration_form']) -> name('auth.registration_form');
+Route::post('/registration',     [AuthController::class,       'registration']) -> name('auth.registration');
+Route::get('/logout',            [AuthController::class,       'logout'])   -> name('auth.logout')      -> middleware('authorized');
 
-Route::get('/blog', [BlogController::class, 'index'])->name('blog.index');
-Route::get('/blog/create', [BlogController::class, 'create'])->name('blog.create');
-Route::post('/blog', [BlogController::class, 'store'])->name('blog.store');
-Route::get('/blog/file_upload', [BlogController::class, 'file_upload'])->name('blog.file_upload');
-Route::post('/blog/file_upload', [BlogController::class, 'file_upload_update'])->name('blog.file_upload_update');
-Route::get('/blog/{id}', [BlogController::class, 'show'])->name('blog.show');
-Route::get('/blog/{id}/edit', [BlogController::class, 'edit'])->name('blog.edit');
-Route::post('/blog/{id}', [BlogController::class, 'update'])->name('blog.update');
-Route::delete('/blog/{id}', [BlogController::class, 'destroy'])->name('blog.destroy');
+Route::get('/spy',               [SpyController::class,        'index']) -> name('spy.index')       -> middleware('admin');
